refactor(db): split queries.js into focused modules by concern (#16)
Pure code motion, no behavior change: every function body is moved
verbatim; only import/export headers differ.

- helpers.js: getGuild, getParticipant, getActiveParticipant,
  dateInGuildTimezone, getOrCreateEntry, isExerciseType
  (previously private, now shared across db modules)
- participation.js: setupGuild, joinChallenge, leaveChallenge
- mutations.js: applyExerciseChange + add/set/adminAdd/adminRemove/
  adminSet wrappers (transaction + SELECT ... FOR UPDATE preserved)
- stats.js: getLeaderboard, getGlobalStats, getUserStats, getUserAllStats
- recap.js: isRecapDue (still in src/db/, not moved elsewhere),
  getDueRecapGuilds, getDailyProgress, markRecapSent
- queries.js: kept as a barrel module re-exporting everything, so the
  command files and scheduler.js keep importing from '../../db/queries.js'

All modules stay under ~200 lines. Import graph is acyclic:
helpers <- participation/mutations/stats/recap <- queries.

Verified with node --check on all files, a live ESM import of the barrel
(all exports incl. EXERCISE_TYPES resolve), and npx eslint .
